WIP: Add patient discharge logic
