+++ controllers/reportController.js
@@ -1,10 +1,6 @@
 const { Report } = require("../models/report");
 const { Patient } = require("../models/patient");
-const {
-  DeleteObjectCommand,
-} = require("@aws-sdk/client-s3");
-const { s3 } = require("../middleware/awsClient");
-const { uploadReport } = require("../middleware/uploadReport");
+const { uploadReport, deleteReport } = require("../middleware/report");
 
 const getReports = async (req, res) => {
   try {
@@ -73,7 +69,7 @@ const getDischargeReports = async (req, res) => {
 const createReport = async (req, res) => {
   try {
     const patientId = req.body.patient;
-    const {name, type} = req.body;
+    const { name, type } = req.body;
     const patient = await Patient.findById({ _id: patientId });
     if (!patient) {
       return res.status(500).json({
@@ -83,7 +79,7 @@ const createReport = async (req, res) => {
 
     const file = req.file;
 
-    const uploadedReportUrl = await uploadReport(patientId, type, name, file)
+    const uploadedReportUrl = await uploadReport(patientId, type, name, file);
 
     res.status(200).json({ success: true, data: uploadedReportUrl });
   } catch (e) {
@@ -124,38 +120,12 @@ const updateReportByReportId = async (req, res) => {
 const deleteReportByReportId = async (req, res) => {
   try {
     const { id } = req.params;
-    const report = await Report.findById(id);
-    if (!report) {
-      return res
-        .status(404)
-        .json({ message: `cannot find any report with ID ${id}` });
+    try {
+      const report = await deleteReport(id);
+      res.status(200).json(report);
+    } catch (error) {
+      res.status(500).json(error)
     }
-    const patientId = await Patient.findOne({ reports: id });
-    const updatedPatient = await Patient.findOneAndUpdate(
-      { _id: patientId },
-      { $pull: { reports: id } },
-      {
-        new: true,
-        runValidators: true,
-      }
-    );
-    if (!updatedPatient) {
-      return res.status(500).json({
-        message: `cannot find any patient tagged to report with ID ${id}`,
-      });
-    }
-
-    const key = report.url.split(".amazonaws.com/")[1];
-    await s3.send(
-      new DeleteObjectCommand({
-        Bucket: "ah-virtual-nursing",
-        Key: key,
-      })
-    );
-
-    await Report.deleteOne({ _id: id });
-
-    res.status(200).json(report);
   } catch (e) {
     res.status(500).json({ message: e.message });
   }
